:lipstick: Split card classes into Stone and Monster
