+++ index.html
@@ -6,7 +6,6 @@
     <link rel="stylesheet" href="./src/stylesheet/main.css">
 </head>
 <body>
-<img src="./src/assets/sprites/game.png" id="game-sprite" style="display:none">
 <script src="./src/javascript/main.js"></script>
 </body>
 </html>
