+++ src/javascript/constructor/board.js
@@ -1,24 +1,22 @@
 // @flow
 
-import { GAME_SPRITE } from '../common/variables';
 import { backgroundTile, boardTiles } from './common/tiles/board-tiles';
 import Canvas from './canvas';
 
-export default class Board {
-  canvas: Canvas;
-
-  constructor(canvas?: Canvas) {
-    this.canvas = canvas || new Canvas('board', GAME_SPRITE);
+export default class Board extends Canvas {
+  constructor() {
+    super({
+      type: 'board',
+    });
     this.drawBackground();
     this.drawBoard();
-    this.canvas.save();
   }
 
   /**
    * Draw background image
    */
   drawBackground() {
-    this.canvas.drawImage(
+    this.drawImage(
       backgroundTile.x,
       backgroundTile.y,
       backgroundTile.width,
@@ -28,7 +26,7 @@ export default class Board {
   }
 
   drawBoard() {
-    this.canvas.drawImage(
+    this.drawImage(
       boardTiles.x,
       boardTiles.y,
       boardTiles.width,
